所得税を別ファイルに分割
